requirements.txt
Django==6.0.1
djangorestframework==3.14.0
sqlparse==0.5.5
asgiref==3.11.0
tzdata==2025.3
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
